# docs/openapi.yaml
components:
...
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
info:
  contact:
    email: support@weather-agent.com
    name: Weather Agent Support
    url: https://weather-agent.com/support
  description: "\n# Weather Agent A2A API\n\nAPI completa para comunicaci\xF3n Agent-to-Agent\
    \ (A2A) del agente meteorol\xF3gico.\n\n## Caracter\xEDsticas\n\n- **JSON-RPC\
    \ 2.0**: Protocolo est\xE1ndar para comunicaci\xF3n entre agentes\n- **Agent Discovery**:\
    \ Descubrimiento autom\xE1tico mediante Agent Card\n- **Capacidades Meteorol\xF3\
    gicas**: Clima actual, pron\xF3sticos y b\xFAsqueda de ubicaciones\n- **Gesti\xF3\
    n de Tareas**: Ejecuci\xF3n as\xEDncrona de operaciones complejas\n- **Monitoreo**:\
    \ Health checks y m\xE9tricas de rendimiento\n\n## Flujo de Comunicaci\xF3n A2A\n\
    \n1. **Descubrimiento**: GET `/.well-known/agent.json`\n2. **Capacidades**: POST\
    \ `/rpc` con m\xE9todo `get_capabilities`\n3. **Ejecuci\xF3n**: POST `/rpc` con\
    \ capacidades espec\xEDficas\n4. **Monitoreo**: GET `/health` y `/status`\n\n\
    ## Autenticaci\xF3n\n\nActualmente no se requiere autenticaci\xF3n. En producci\xF3\
    n se recomienda:\n- API Keys para agentes registrados\n- OAuth 2.0 para aplicaciones\
    \ cliente\n- mTLS para comunicaci\xF3n segura entre agentes\n\n## Rate Limiting\n\
    \n- **Desarrollo**: Sin l\xEDmites\n- **Producci\xF3n**: 100 requests/minuto por\
    \ agente\n        "
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Weather Agent A2A API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /tasks/{task_id}:
    get:
      description: "Obtener estado de una tarea espec\xEDfica."
      operationId: get_task_tasks__task_id__get
      parameters:
      - in: path
        name: task_id
        required: true
        schema:
          title: Task Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Task
servers:
...